--- .coveragerc
@@ -1,2 +1,0 @@
-[run]
-omit = *test*, ./Entities/*, __init__.py, db.py
