+++ .gitignore
@@ -4,6 +4,3 @@ token.txt
 .env
 __pycache__
 venv
-.pytest_cache/
-*.session
-*.session-journal
