+++ Entities/Member.py
@@ -1,0 +1,6 @@
+
+class Member:
+    def __init__(self, id: int, name: str, tgNum: int):
+        self.id: int = id
+        self.name: str = name
+        self.tgNum: int = tgNum
