--- RunSystemTests.bat
@@ -1,2 +1,0 @@
-python -m pytest -m system
-pause
