--- RunUnitTests.bat
@@ -1,2 +1,0 @@
-python -m pytest -m unit
-pause
