--- Src/Entities/Subject.py
@@ -1,5 +1,0 @@
-
-class Subject():
-    def __init__(self, id: int = 0, title: str = ''):
-        self.id: int = id
-        self.title: str = title
