+++ db.py
@@ -1,0 +1,28 @@
+from typing import List
+import psycopg2
+from dotenv import load_dotenv
+import os
+from datetime import datetime
+
+from Entities.Member import Member
+from Entities.Subject import Subject
+from Services.MemberService import MemberService
+
+class Database:
+    def __init__(self):
+        load_dotenv()
+        self.connection = psycopg2.connect(
+            host=os.getenv('host'),
+            user=os.getenv('user'),
+            password=os.getenv('password'),
+            database=os.getenv('db_name'),
+            port=os.getenv('port')
+        )
+        self.connection.autocommit = True
+        with self.connection.cursor() as cur:
+            cur.execute("select version();")
+            print(f"server vers:  {cur.fetchone()}")
+
+    def close(self):
+        self.connection.close()
+        print("[INFO] Close connection with DB")
