--- pytest.ini
@@ -1,8 +1,0 @@
-[pytest]
-testpaths = Tests
-addopts = --strict-markers
-pythonpath = Src
-markers =
-    unit
-    integration
-    system
